secure_assess-core/requirements/base.txt
# Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0

# Database
sqlalchemy>=1.4.0
alembic>=1.7.0
asyncpg>=0.24.0
psycopg2-binary>=2.9.0
redis>=4.0.0

# Async Tools
celery>=5.2.0
aiohttp>=3.8.0

# Security Tools
bandit>=1.7.0
semgrep>=1.5.0
dependency-check>=0.5.0
python-owasp-zap-v2.4>=0.4.0

# Git Integration
gitpython>=3.1.0

# Utility
pydantic>=1.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.5
requests>=2.26.0
PyYAML>=6.0.0
